==> scripts/03_merging_data.ipynb <==
<h1><center> PPOL 5203 : Final Project  <br><br> 
<font color='grey'> Merging Data <br><br>
Ibadat Jarg and Helen Wang</center></center> <h1> 

#setting library
import pandas as pd
import numpy as np

#reading in files
op_traffic = pd.read_csv("../data/cleaned_data/opened_cb_traffic.csv")
unop_traffic = pd.read_csv("../data/cleaned_data/unopened_cb_traffic.csv")

#add indicator cols to df to show whether they're unopened or opened bikeshares for later merging
op_traffic["opened_station"] = 1
unop_traffic["opened_station"] = 0

#merging df of opened capital bikeshares and unopened capital bikeshares
merged_df = pd.concat([op_traffic, unop_traffic], ignore_index = True)

#add indicator col for if opened in 2013
merged_df["opened_2013"] = np.where(merged_df["open_year"] == 2013.0, True, False)

#checking snap shot of data
merged_df.iloc[2:5]

#saving final data
merged_df.to_csv("../data/cleaned_data/final_data.csv", index = False)
