==> scripts/06_proposed_stations.ipynb <==
<h1><center> PPOL 5203 : Final Project  <br><br> 
<font color='grey'> Proposed Stations <br><br>
Ibadat Jarg and Helen Wang</center></center> <h1> 

#set libraries
import pandas as pd
import numpy as np
import geopandas as gpd
from plotnine import *

#read files
df = pd.read_csv("../data/cleaned_data/final_data.csv")
gdf_points = gpd.GeoDataFrame(df, geometry=gpd.points_from_xy(df['long'], df['lat']), crs="EPSG:4326")

#unopened stations
unopened_stations = df[df["opened_station"] == 0]

#sorting it by highest traffic in 2019
unopened_stations = unopened_stations.sort_values(by="2019", ascending=False)

#identifying top 25%
unopened_stations["cutoff"] = np.where(unopened_stations["2019"] >= unopened_stations["2019"].describe()['75%'], True, False)

#assessing distribution
unopened_stations["cutoff"].value_counts()

#reading in geographic data
dc = gpd.read_file('../data/raw_data/dc_shape_file_zip') #dc census tracts

#plot of proposed, unopened bikeshares 
map2 = (
ggplot(data = unopened_stations) +
    geom_map(data = dc,
            mapping = aes(geometry = 'geometry'),
            color = "white",
            alpha = 0.5) +
    geom_point(mapping = aes(x = 'long', y = 'lat', color = "2019", shape = "cutoff", size = "2019")
        ) +
    labs(title = "Proposed Unopened Bikeshare Locations by Traffic", color = "AADT", shape = "Top 25%") +
    theme_void() +
    theme(
      plot_title=element_text(face="bold")
    ) +
    scale_color_distiller(type = "div",palette="RdYlGn", direction = 1)  +
    guides(size = False)
)

map2

map2.save('../visualizations/graphs/map2.png', dpi=300)

unopened_stations[unopened_stations["cutoff"] == True].filter(["name", "long", "lat", "2019"]).reset_index()
